lmsWebAPI/requirements.txt
asgiref==3.8.1
channels==4.2.2
Django==5.2.1
django-allauth==65.8.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
setuptools==78.1.1
sqlparse==0.5.3
wheel==0.45.1
whitenoise==6.9.0
